requirements.txt
Django==4.2.20
psycopg2-binary==2.9.10

python-telegram-bot==20.8
python-dotenv==1.0.1

pytest==8.3.5
pytest-django==4.11.1
